Updated templates + a refactor
