+++ lib/emailTemplates/inviteToForms.ts
@@ -1,10 +1,10 @@
 /**
  * Send an invitation to register for GC Forms
  *
- * @param senderName
- * @param message
- * @param registerUrlEn
- * @param registerUrlFr
+ * @param senderName Sender's name
+ * @param message Message
+ * @param registerUrlEn Registration URL
+ * @param registerUrlFr Registration URL
  * @returns
  */
 export const inviteToForms = (
@@ -23,7 +23,9 @@ ${message}
 
 ===
 
-French version
+${senderName} has invited you to create a GC Forms account.
+
+${message}
 
 [Create account](${registerUrlFr})
 `;
