+++ lib/emailTemplates/ownerRemovedNotification.ts
@@ -1,22 +1,23 @@
 /**
  * Notify owners that an owner has been added to a form.
  *
- * @param formTitle
- * @param formOwner
+ * @param formTitle Form title
+ * @param formOwner Removed owner's name
  * @returns
  */
 export const ownerRemovedNotification = (formTitle: string, formOwner: string): string => `
 (la version française suit)
 
-Hello ${formOwner},
-
-You have been removed from the form ${formTitle}:
+${formOwner} been removed from the form ${formTitle}:
 
 Thanks,
 The GC Forms team
 
 ======
 
-french version
+${formOwner} been removed from the form ${formTitle}:
+
+Thanks,
+The GC Forms team
 
 `;
